pyproject.toml
[project]
name = "Photo-Composition-Designer"
dynamic = ["version"]
description = "Feature-rich Python project template for Photo-Composition-Designer."
authors = [
    {name = "pamagister"}
]
readme = "docs/index.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "astral>=3.2",
    "config-cli-gui>=0.3.0",
    "exifread>=3.5.1",
    "geopandas>=1.1.1",
    "holidays>=0.84",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "pytz>=2025.2",
    "shapely>=2.1.1",
]

# Dev dependencies as optional dependencies
[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-mock>=3.14.1",
    "coverage>=7.8.2",
    "flake8>=7.2.0",
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest-cov>=6.1.1",
    "mypy>=1.16.0",
    "gitchangelog>=3.0.4",
    "pyinstaller>=5.8",
    "pre-commit>=4.2.0",
    "ruff>=0.11.13"
]

docs = [
    "mkdocs>=1.6.1",
    "mkdocs-awesome-nav>=2.6.1",
    "pygments>=2.19.1",
]

[project.scripts]
Photo-Composition-Designer = "Photo_Composition_Designer.cli.cli:main"
Photo-Composition-Designer-gui = "Photo_Composition_Designer.gui.gui:main"

[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/Photo_Composition_Designer/_version.py"

[tool.deptry]
pep621_dev_dependency_groups = ["dev", "docs"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["src/Photo_Composition_Designer/_version.py"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E721", "UP015", "B009", "B905"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
